+++ src/jvmTest/kotlin/kotools/types/collections/NotEmptySetTest.kt
@@ -1,18 +1,28 @@
 package kotools.types.collections
 
 import kotools.assert.*
+import kotools.types.*
 import kotools.types.core.RandomValueHolder
-import kotools.types.pairBy
-import kotools.types.runMapFirst
-import kotools.types.runPairBy
 
 class NotEmptySetTest : RandomValueHolder {
     // ---------- Builders ----------
 
     @Test
+    fun notEmptySetOf_should_pass(): Unit =
+        (randomInt to arrayOf(randomInt, randomInt))
+            .runPairBy { notEmptySetOf(first, *second) }
+            .runMapSecond { setOf(first) + second }
+            .run {
+                first.size assertEquals second.size
+                first.forEachIndexed { index: Int, element: Int ->
+                    element assertEquals second.elementAt(index)
+                }
+            }
+
+    @Test
     fun constructor_should_pass(): Unit = setOf(randomInt, randomInt, randomInt)
         .runPairBy { first() to toList().subList(1, size).toTypedArray() }
-        .runMapFirst { NotEmptySet(first, *second) }
+        .runMapFirst { notEmptySetOf(first, *second) }
         .run {
             first.size assertEquals second.size
             first.forEachIndexed { index: Int, element: Int ->
@@ -64,7 +74,7 @@ class NotEmptySetTest : RandomValueHolder {
 
     @Test
     fun collection_toNotEmptySetOrElse_should_pass_with_a_not_empty_collection(): Unit =
-        (setOf(randomInt, randomInt) to NotEmptySet(randomInt))
+        (setOf(randomInt, randomInt) to notEmptySetOf(randomInt))
             .runPairBy { first.toNotEmptySetOrElse { second } }
             .run {
                 first.size assertEquals second.first.size
@@ -76,7 +86,7 @@ class NotEmptySetTest : RandomValueHolder {
 
     @Test
     fun collection_toNotEmptySetOrElse_should_return_the_default_value_with_an_empty_collection(): Unit =
-        (emptySet<Int>() to NotEmptySet(randomInt, randomInt))
+        (emptySet<Int>() to notEmptySetOf(randomInt, randomInt))
             .runPairBy { first.toNotEmptySetOrElse { second } }
             .run {
                 first.size assertNotEquals second.first.size
@@ -88,7 +98,7 @@ class NotEmptySetTest : RandomValueHolder {
 
     @Test
     fun array_toNotEmptySetOrElse_should_pass_with_a_not_empty_array(): Unit =
-        (arrayOf(randomInt, randomInt) to NotEmptySet(randomInt))
+        (arrayOf(randomInt, randomInt) to notEmptySetOf(randomInt))
             .runPairBy { first.toNotEmptySetOrElse { second } }
             .run {
                 first.size assertEquals second.first.size
@@ -100,7 +110,7 @@ class NotEmptySetTest : RandomValueHolder {
 
     @Test
     fun array_toNotEmptySetOrElse_should_return_the_default_value_with_an_empty_array(): Unit =
-        (emptyArray<Int>() to NotEmptySet(randomInt, randomInt))
+        (emptyArray<Int>() to notEmptySetOf(randomInt, randomInt))
             .runPairBy { first.toNotEmptySetOrElse { second } }
             .run {
                 first.size assertNotEquals second.first.size